make less dependent on import order
